Refactor: DNSWorld
